requirements.txt
Django>=5.0
strawberry-graphql[django]>=0.246
